Move ScopedLocalStorage into walletlink directory
