+++ packages/wallet-sdk/src/sign/walletlink/storage/ScopedLocalStorage.ts
